requirements.txt
flask
requests
sqlitecloud
python-dotenv
PyJWT
flask-restx